Move navigation between pages to the sidebar
It was in the menu-dropdown before, which is kinda hard to access,
considering that the main sections should be easy to hop around.

Also make sidebar hidden by default for now, because it has a more
polished look this way. As I'll add project overviews, I'll consider
changing it back.

Currently the navigation between projects/filters on the todolist page
is broken, because it's part of a higher route. The next commit will
address this issue.