Add the first versions of heatmap for comparison
While I was thinking of the ways to adequately introduce the heatmap for
eliminating bad habits, I realized that my implementation of CellPeriod
has pitfalls - they can't start on the same week, otherwise it breaks
everything. CellPeriod has height of the container, so it must always be
put between two adjacent weeks, not on the same one.

So the Cell I had in first versions would come in handy here - for
comparison I made a TimelineSimple and used the heatmap-data generation
which I haven't touched since making CellPeriod component.

Also I realized that I've shot myself in the leg by not using date-fns
from the start. I've made some changes to the TimelineSimple with using
its features, but decided to do the rest in the next commit.