Add basic redux store with RTK query for todolist
Later I'll be extending the functionality of the store so that addition,
deletion, and modification of tasks is also supported.

Redux was necessary for keeping the data shared between multiple
projects/pages. But I also discovered RTK query, which is pretty lit for
displaying spinners while data is still fetching.

I also decreased the amount of api requests. I realized that filtering
data on the server is not really rational when I'm not even sure how
much users I'll have. And it's just easier to manage client-wise.