Add optimistic updates for task deletion/creation
I thought that this wasn't possible, but it is - I just assumed that it
wasn't because it didn't work.

Anyway, the problem was that I didn't fullfill a promise. The async
function declaration in createTaskMutation was out of place - it was
modifying a copy of cache, not the cache itself.

You might ask, why do you need to go to such long ways as to wait for a
request to fullfill? It's simple. I structured my app in such a way that
taskID is necessary, and I when I create tasks, I assign them on the
server, not on the client. Hence the need for the promise.

...I also learned about upsertQueryData. It turned out to be a very
recent PR, but after reading the docs, I realized that it's not really
doing what I need.