Add hooks for common useEffect/useState logic
I noticed that I often rely on the useEffect which changes the state
once the component has loaded, so that I can hide the spinner and show
the page. So I created a hook which does that.

Also, I noticed that the logic in the App.jsx which supposedly should've
hidden the spinner, did nothing, because the spinner was deleted
automatically on mount.

The lines:
<div class="root" />
<div class="loader" />

Were rendered as:
<div class="root">
  <div class="loader"></div>
</div>

So there was nothing to hide in the first place. I cleaned the html file
to use proper nesting.