Make cell colors/border color depend on theme
Also, in the future I'll need to do something about the color schemes of
cells. Either generate them, or keep it the same way where I calculate
it and use hex color opacity. Border colors might also utilize something
like that for more contrast of the cell that is pointed at.