Change habit tips to be a dynamic div
I faced some frustration while handling the fact that it's impoosible to
set overflow-x: hidden and overflow-y: visible at the same time. I
decided to handle it the way github does - just add a div that would
calculate its position dynamically when you hover on cells.

Currently, when you hover on the :before part of the cell-period, the
tip will display just a little to the top of the :before element. I
think it would be worth it to display it beneath the :before element,
although I'm not sure whether it will or won't cause the user to get
lost.

Anyway, I made a lot of changes in this commit, like handling offsets to
prevent overflow of window width properly (although you can improve the
formula a little bit - I made the tip width mutable depending on the
amount of text with nowrap, but it still uses previous half-lengths of
100px when calculating offsets).