Make cell colors mix with the default cell color
Before, I used rgba for showing different weight of cells. But it seems
that the recent CSS update made the rgba colors work the same way as
opacity. So the borders of cell periods stutter a little bit because of
that. So now I've got no choice but compute the cell colors through JS.

Well, it's not like I didn't need to do that - the colors had to start
from some base, because the default cell color was brighter than the
cells with alpha > 0. They didn't fit the picture.

But now the whole page with habits needs to stutter for a couple of
seconds when I change the theme color. It's a small price to pay. I
thought about setting the regular cells with opacity, but forgot that
this way the previous paragraph pretty much doesn't get fixed. Or I need
to add one more cell as a background, so that I can work with opacity.
But that'll make the loading of those pages even longer, so that's a
no-go.