Fix heatmaps and project deleting improperly
It was kind of annoying to test heatmaps when there's that big of a hole
in the system. I had to fix it asap to go any further. To address the
elephant in the room.

The bug was that when you created a new project, you had to reload the
page in order to see the heatmaps. And the tasklists aren't deleted when
the project is deleted.. I fixed that by allowing the tasks have no
linked projects. I believe it's reasonalbe to add an option to choose
whether you want to delete all the tasks from the project when deleting
it, or keep them.