Calculate heatmap width based on the content
Also, I realized that one of the commits on this branch was kinda
unnecessary - the one where I've set offsets of 12.3 and 12.35 to
cell-periods. It was caused by the fact that I used flexbox, and it was
kinda obvious that it would behave this way. I've just reset it to the
previous values. I'm gonna pretend I intended this from the start xdd.