Add cell-period-connector to fix cell-period width
Previously, the cell-periods used only the parent cell-period to connect
&-before and &-after, but since I started using the amount of weeks in
order to calculate the heatmap width, it was no longer an option.

The reason is that if I used a wide cell-period parent in order to
connect before and after elements while cell-period width is 0, then the
width of such cell-period would exceed what it normally takes for two
columns.

Hence, I added a connector between after and before elements to fix the
width.