Cells with multiple actions (fractured cells)
I finally started working on the feature that I always had in mind, that
I thought would make neohabit pretty different from other apps. The
first implementation kinda sucks, there's a lot of things that I need to
do before I can consider it a full feature).

For example, I need to create something like goals for a day (just the
amount of times the habit is expected to be done (or initiated) in a
day).

Also there should be a way of showing that a habit has more than 9 or 16
actions performed. I implemented it up to 16, but it looks kinda messy,
so I decided to limit the amount of cell-fractions to 9. But still it's
cool.

Another idea for fractured cells is to have a setting to display cells
with smaller fractions as numbers (like 1, 2, 3... or I, II, III... or
ー、ニ、三... or as parts of 正)...

fml