Make proper target changes and handle other cases
Phew, it's a weighty component. All the cases have to be handled very
carefully. Now if the target starts before the dateStart, it's still
handled properly. The start of the target is respected, as well as the
case where the data point which changes the target period is before the
dateStart.